+++ src/main/java/gift/config/WebConfig.java
@@ -1,4 +1,4 @@
-package gift;
+package gift.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
